Create boilerplate.css
Students: as I discussed in class, CSS boilerplate exists to normalize the differences in the ways that browsers display content. All the intimidating-looking gobbledegook in this file exists just so that what you see on Chrome looks the same as what your mom sees when she looks at your new web page in IE. 

And if she's using IE6, please help her install a newer web browser that isn't so vulnerable to security risks.